+++ Tests/Runtime/BaselibNetworkInterfaceTests.cs
@@ -1,7 +1,6 @@
 using System.Threading;
 using NUnit.Framework;
 using Unity.Networking.Transport;
-using Unity.Networking.Transport.Utilities;
 using UnityEngine;
 using UnityEngine.TestTools;
 using System.Linq;
@@ -49,9 +48,15 @@ namespace Unity.Networking.Transport.Tests
             }
         }
 
+        private void FakeSocketFailure(BaselibNetworkInterface baselibInterface)
+        {
+            var baselib = baselibInterface.m_Baselib[0];
+            baselib.m_SocketStatus = BaselibNetworkInterface.SocketStatus.SocketNeedsRecreate;
+            baselibInterface.m_Baselib[0] = baselib;
+        }
+
         [Test]
-        [UnityPlatform(include = new[] { RuntimePlatform.IPhonePlayer })]
-        public void Baselib_AfterAppSuspension_SocketIsRecreated()
+        public void Baselib_AfterSocketFailure_SocketIsRecreated()
         {
             using (var baselibInterface = new BaselibNetworkInterface())
             using (var dummyDriver = NetworkDriver.Create())
@@ -69,34 +74,75 @@ namespace Unity.Networking.Transport.Tests
                 packetReceiver.m_Driver = dummyDriver;
                 baselibInterface.ScheduleReceive(packetReceiver, default).Complete();
 
-                // Fake an app suspension by manually calling the focus callback. We add sleeps
-                // around the call to ensure the timestamp is different from the receice jobs.
-                Thread.Sleep(5);
-                AppForegroundTracker.OnFocusChanged(true);
-                Thread.Sleep(5);
+                // Sleep to ensure different update times.
+                Thread.Sleep(2);
+
+                FakeSocketFailure(baselibInterface);
 
                 dummyDriver.ScheduleUpdate().Complete();
                 packetReceiver.m_Driver = dummyDriver;
                 baselibInterface.ScheduleReceive(packetReceiver, default).Complete();
 
                 Assert.AreNotEqual(socket, baselibInterface.m_Baselib[0].m_Socket);
+
+                LogAssert.Expect(LogType.Warning, "Socket error encountered; attempting recovery by creating a new one.");
             }
         }
 
         [Test]
-        [UnityPlatform(include = new[] { RuntimePlatform.IPhonePlayer })]
-        public void Baselib_AfterAppSuspension_CanSendReceive()
+        public void Baselib_AfterBackToBackSocketFailures_SocketIsFailed()
+        {
+            using (var baselibInterface = new BaselibNetworkInterface())
+            using (var dummyDriver = NetworkDriver.Create())
+            {
+                var settings = new NetworkSettings();
+                baselibInterface.Initialize(settings);
+                baselibInterface.CreateInterfaceEndPoint(NetworkEndPoint.AnyIpv4, out var endpoint);
+                Assert.Zero(baselibInterface.Bind(endpoint));
+
+                var packetReceiver = new NetworkPacketReceiver();
+
+                dummyDriver.ScheduleUpdate().Complete();
+                packetReceiver.m_Driver = dummyDriver;
+                baselibInterface.ScheduleReceive(packetReceiver, default).Complete();
+
+                // Sleep to ensure different update times.
+                Thread.Sleep(2);
+
+                FakeSocketFailure(baselibInterface);
+
+                dummyDriver.ScheduleUpdate().Complete();
+                packetReceiver.m_Driver = dummyDriver;
+                baselibInterface.ScheduleReceive(packetReceiver, default).Complete();
+
+                LogAssert.Expect(LogType.Warning, "Socket error encountered; attempting recovery by creating a new one.");
+
+                // Sleep to ensure different update times.
+                Thread.Sleep(2);
+
+                FakeSocketFailure(baselibInterface);
+
+                dummyDriver.ScheduleUpdate().Complete();
+                packetReceiver.m_Driver = dummyDriver;
+                baselibInterface.ScheduleReceive(packetReceiver, default).Complete();
+
+                Assert.AreEqual((int)Error.StatusCode.NetworkSocketError, dummyDriver.ReceiveErrorCode);
+
+                LogAssert.Expect(LogType.Error, "Unrecoverable socket failure. An unknown condition is preventing the application from reliably creating sockets.");
+                LogAssert.Expect(LogType.Error, "Error on receive, errorCode = -10");
+            }
+        }
+
+        [Test]
+        public void Baselib_AfterSocketRecreation_CanSendReceive()
         {
             using (var server = NetworkDriver.Create())
             using (var client = NetworkDriver.Create())
             {
                 ConnectServerAndClient(NetworkEndPoint.LoopbackIpv4, server, client, out _, out var connection);
 
-                // Fake an app suspension by manually calling the focus callback. We add sleeps
-                // around the call to ensure the timestamp is different from the driver updates.
-                Thread.Sleep(5);
-                AppForegroundTracker.OnFocusChanged(true);
-                Thread.Sleep(5);
+                var clientBaselibInterface = (BaselibNetworkInterface)client.NetworkInterface;
+                FakeSocketFailure(clientBaselibInterface);
 
                 // Let the server and client recreate their sockets.
                 client.ScheduleUpdate().Complete();